ingest_to_bronze/pyproject.toml
[project]
name = "mvp"
version = "0.0.1"
authors = [{ name = "ceberiloclaus@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    # Any dependencies for jobs and pipelines in this project can be added here
    # See also https://docs.databricks.com/dev-tools/bundles/library-dependencies
    #
    # LIMITATION: for pipelines, dependencies are cached during development;
    # add dependencies to the 'environment' section of your pipeline.yml file instead
    "click>=8.3.1",
    "databricks-sdk>=0.32.0",
]

[dependency-groups]
dev = [
    "databricks-connect==15.4",
    "pytest>=7.0.0",
]

[project.scripts]
ingest_to_bronze = "mvp.transformations.ingest_to_bronze:run_main"
table_summarization = "mvp.summarization.table_summarization:run_main"
create_dashboard = "mvp.summarization.create_dashboard:run_main"

transform_default_product_info_to_silver = "mvp.transformations.default_product_info_silver:run_main"
transform_product_specifications_silver = "mvp.transformations.product_specifications_silver:run_main"
transform_recommended_metadata_silver = "mvp.transformations.recommended_metadata_silver:run_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 125

[tool.hatch.build.targets.wheel]
packages = ["src/mvp", "src/shared"]
